chore: split up bridge files
